Removed StateManager, Reworked Command "State", Updated Structures, Temporary File Encryption
Reworked the "StateManager" system to not be as complicated. Commands just now directly have a state controller on them. This ensures that certain commands cannot be used in other places.

Additionally, refactored all the code to reflect these changes.

While doing this, I finished the base updates for all of the Structure.

Next Goal:

Player Creation.